Crate embassy_macros

embassy-macros

An Embassy project.

Macros for creating the main entry point and tasks that can be spawned by embassy-executor.

NOTE: The macros are re-exported by the embassy-executor crate which should be used instead of adding a direct dependency on the embassy-macros crate.

Minimum supported Rust version (MSRV)

The task and main macros require the type alias impl trait (TAIT) nightly feature in order to compile.

License

This work is licensed under either of

Apache License, Version 2.0 (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0)

MIT license (LICENSE-MIT or http://opensource.org/licenses/MIT)

at your option.

Attribute Macros

main_cortex_m

Creates a new executor instance and declares an application entry point for Cortex-M spawning the corresponding function body as an async task.

main_riscv

Creates a new executor instance and declares an application entry point for RISC-V spawning the corresponding function body as an async task.

main_std

Creates a new executor instance and declares an application entry point for STD spawning the corresponding function body as an async task.

main_wasm

Creates a new executor instance and declares an application entry point for WASM spawning the corresponding function body as an async task.

task

Declares an async task that can be run by embassy-executor. The optional pool_size parameter can be used to specify how many concurrent tasks can be spawned (default is 1) for the function.